fix: restructured component
