+++ .github/workflows/release-please.yml
@@ -28,5 +28,3 @@ jobs:
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
         if: ${{ steps.release.outputs.release_created }}
-      - run: npm run deploy
-        if: ${{ steps.release.outputs.release_created }}
