+++ .storybook/preview-head.html
@@ -2,6 +2,10 @@
   href="https://fonts.googleapis.com/css?family=Material+Icons&display=block"
   rel="stylesheet"
 />
+<link
+  href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@300&family=Roboto:wght@300;400;500&display=swap"
+  rel="stylesheet"
+/>
 <style>
   body {
     --oscd-action-pane-theme-on-surface: #657b83;
