+++ README.md
@@ -12,7 +12,7 @@ npm i @openscd/oscd-action-pane
 
 ```html
 <script type="module">
-  import 'oscd-action-pane';
+  import '@openscd/oscd-action-pane';
 </script>
 
 <oscd-action-pane></oscd-action-pane>
@@ -60,12 +60,23 @@ To build a production version of Storybook, run
 npm run storybook:build
 ```
 
-
 ## Tooling configs
 
 For most of the tools, the configuration is in the `package.json` to reduce the amount of files in your project.
 
 If you customize the configuration a lot, you can consider moving them to individual files.
+
+### Icon font
+
+Material Icons are being used for the icons. This font needs to be added in the html first.
+You can add it like so:
+
+```html
+<link
+  href="https://fonts.googleapis.com/css?family=Material+Icons&display=block"
+  rel="stylesheet"
+/>
+```
 
 ## Local Demo with `web-dev-server`
 
@@ -75,9 +86,10 @@ npm start
 
 To run a local development server that serves the basic demo located in `demo/index.html`
 
-## `oscd-action-pane.ts`:
 
-### class: `OscdActionPane`
+## `src/oscd-action-pane.ts`:
+
+### class: `OscdActionPane`, `oscd-action-pane`
 
 #### Superclass
 
@@ -85,21 +97,52 @@ To run a local development server that serves the basic demo located in `demo/in
 | ------------ | ------ | ------- |
 | `LitElement` |        | lit     |
 
-#### Properties/Attributes
+#### Fields
 
-| Name          | Type      | Default     | Description                                                    |
-| ------------- | --------- | ----------- | -------------------------------------------------------------- |
-| `label`       | `string`  | `undefined` | Caption text, displayed in the header.                         |
-| `icon`        | `string`  | `undefined` | Icon name, displays icon as described in `@material/mwc-icon`. |
-| `secondary`   | `boolean` | `false`     | Color header with secondary theme color while focus is within. |
-| `highlighted` | `boolean` | `false`     | highlight pane with dotted outline item.                       |
-| `level`       | `number`  | `1`         | nesting level, default (closest pane ancestor's level) + 1     |
+| Name          | Privacy | Type                  | Default | Description                                                   | Inherited From |
+| ------------- | ------- | --------------------- | ------- | ------------------------------------------------------------- | -------------- |
+| `label`       |         | `string \| undefined` |         | caption text, displayed in the header                         |                |
+| `icon`        |         | `string \| undefined` |         | icon name, displayed unless the "icon" slot is filled         |                |
+| `secondary`   |         | `boolean`             | `false` | color header with secondary theme color while focus is within |                |
+| `highlighted` |         | `boolean`             | `false` | highlight pane with dotted outline                            |                |
+| `level`       |         | `number`              | `1`     | nesting level, default (closest pane ancestor's level) + 1    |                |
+
+#### CSS Properties
+
+| Name                                  | Default                   | Description                        |
+| ------------------------------------- | ------------------------- | ---------------------------------- |
+| `--oscd-action-icon-theme-primary`    | `--oscd-theme-primary`    | Color for border on even levels.   |
+| `--oscd-action-icon-theme-on-primary` | `--oscd-theme-on-primary` | Pane color for the uneven levels.  |
+| `--oscd-action-icon-theme-secondary`  | `--oscd-theme-secondary`  | Color for border on uneven levels. |
+| `--oscd-action-pane-theme-surface`    | `--oscd-theme-surface`    | Pane color for the even levels.    |
+| `--oscd-action-icon-theme-on-surface` | `--oscd-theme-on-surface` | Icon and label color.              |
+| `--oscd-action-icon-theme-font`       | `--oscd-theme-font`       | Font for label.                    |
+
+#### Slots
+
+| Name     | Description                                                              |
+| -------- | ------------------------------------------------------------------------ |
+| `action` | May contain up to eight icon buttons.                                    |
+| `icon`   | If filled overrides the icon property.                                   |
+|          | The default slot will be rendered into the pane body in a single column. |
+
 <hr/>
 
 ### Exports
 
-| Kind | Name               | Declaration      | Module              | Package |
-| ---- | ------------------ | ---------------- | ------------------- | ------- |
-| `js` | `oscd-action-pane` | oscd-action-pane | oscd-action-pane.ts |         |
+| Kind | Name             | Declaration    | Module                  | Package |
+| ---- | ---------------- | -------------- | ----------------------- | ------- |
+| `js` | `OscdActionPane` | OscdActionPane | src/oscd-action-pane.ts |         |
+
+## `src/OscdActionPane.ts`:
+
+### Exports
+
+| Kind                        | Name               | Declaration    | Module                   | Package |
+| --------------------------- | ------------------ | -------------- | ------------------------ | ------- |
+| `custom-element-definition` | `oscd-action-pane` | OscdActionPane | /src/oscd-action-pane.js |         |
+| `js`                        | `OscdActionPane`   | OscdActionPane | src/OscdActionPane.ts    |         |
+
+
 
 &copy; 2023 Alliander N.V.
