+++ custom-elements-manifest.config.js
@@ -1,13 +1,14 @@
 import { readmePlugin } from 'cem-plugin-readme';
 
 export default {
-  globs: ['*.ts'],
-  exclude: ['**/*.spec.ts', '**/*.test.ts', '**/*.stories.ts'],
+  globs: ['src/*.ts'],
+  exclude: ['test/*.ts', 'stories/*.ts'],
   litelement: true,
   plugins: [
     readmePlugin({
-      header: 'README.head.md',
-      footer: 'README.foot.md',
+      header: 'head.md',
+      footer: 'foot.md',
+      private: 'hidden',
     }),
-  ]
-}
+  ],
+};
