+++ demo/index.html
@@ -6,6 +6,10 @@
       href="https://fonts.googleapis.com/css?family=Material+Icons&display=block"
       rel="stylesheet"
     />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@300&family=Roboto:wght@300;400;500&display=swap"
+      rel="stylesheet"
+    />
     <style>
       body {
         --oscd-action-pane-theme-on-surface: #657b83;
@@ -23,7 +27,7 @@
 
     <script type="module">
       import { html, render, css } from 'lit';
-      import '../dist/src/oscd-action-pane.js';
+      import '../dist/OscdActionPane.js';
       import '@material/mwc-icon';
 
       const title = 'Click this pane!';
