+++ tsconfig.json
@@ -14,9 +14,9 @@
     "outDir": "dist",
     "sourceMap": true,
     "inlineSources": true,
-    "rootDir": "./",
+    "rootDir": "./src",
     "declaration": true,
     "incremental": true
   },
-  "include": ["**/*.ts"]
+  "include": ["src/*.ts"]
 }
