+++ tsconfig.test.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "rootDir": "./"
+  },
+  "extends": "./tsconfig",
+  "include": ["*/*.ts"]
+}
